+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog 
 
+### Version 2.1.0
+
+- For NLog and Log4Net when exception is traced with a custom message, custom message is added to the properties collection and reported to ApplicationInsights.
+- Update Application Insights API reference to 2.1.0 (TODO: add reference to release page)
+
 ### Version 2.0.0
 
 - Update Application Insights API reference to [2.0.0](https://github.com/Microsoft/ApplicationInsights-dotnet/releases/tag/v2.0.0)